requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
pydantic-settings>=2.2.1
langchain>=0.1.9
langchain-community>=0.0.24
faiss-cpu
rank-bm25>=0.2.2
sentence-transformers>=2.5.1
pypdf2>=3.0.1
python-docx>=1.1.0
spacy>=3.7.4
requests>=2.31.0
python-multipart>=0.0.9
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.1.2
streamlit>=1.31.1
tiktoken>=0.6.0
memory-profiler>=0.61.0
reportlab>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
langsmith>=0.1.0
datasets>=2.14.0
rouge_score>=0.1.2
huggingface_hub>=0.20.0
